+++ CHANGES.md
@@ -1,7 +1,12 @@
 Change list
 ===========
 ####22/05/2013
-1. Fixed hardcoded urls in $.pimpMyFilter().
+1. Fix of hardcoded urls in `$.pimpMyFilter()`.
+2. Bootstrap `typeahead` fix.
+3. Added simple support of `ManyToManyField`.
+4. Added `limit` to settings. Now you can limit num of conditions for user. Default `limit:0` (unlimited).
+5. Updated fields declaration.
+6. Added regexp html5 validation to fields, using `pattern` attribute. Not tested yet.
 
 ####21/05/2013
 1. Init